Simplify request-content internals
