+++ projects/egress-gate/src/egress_gate/request_content/json.py
@@ -3,9 +3,6 @@
 from __future__ import annotations
 
 import json
-import re
-from dataclasses import dataclass
-from enum import StrEnum
 from typing import Annotated, Literal, TypeAlias
 
 from pydantic import Field, field_validator
@@ -13,14 +10,19 @@ from pydantic import Field, field_validator
 from egress_gate.base import StrictDomainModel
 from egress_gate.constants import (
     MAX_BODY_BYTES,
-    MAX_JSON_DEPTH,
-    MAX_JSON_NODES,
     MAX_JSON_SELECTED_NODES,
     MAX_JSON_SELECTED_TEXT_BYTES,
     MAX_JSON_SELECTOR_SEGMENTS,
     MAX_JSON_SELECTORS,
 )
 from egress_gate.errors import BodyFormatError, GateInputError, GateLimitExceededError
+from egress_gate.request_content._json_parser import _JsonNode, _JsonParser
+from egress_gate.request_content.json_values import (
+    JsonNode,
+    JsonNodeKind,
+    JsonTextNode,
+)
+from egress_gate.request_content.text import TextReplacement
 from egress_gate.string_validators import ScalarString
 from egress_gate.timeout import Timeout
 
@@ -65,33 +67,6 @@ class JsonSelector(StrictDomainModel):
         if isinstance(value, list | tuple):
             return tuple(value)
         return value
-
-
-class JsonNodeKind(StrEnum):
-    """The JSON value kind at one immutable document node."""
-
-    OBJECT = "object"
-    ARRAY = "array"
-    STRING = "string"
-    NUMBER = "number"
-    BOOLEAN = "boolean"
-    NULL = "null"
-
-
-class JsonNode(StrictDomainModel):
-    """Opaque immutable reference to one node in a ``JsonDocument``."""
-
-    id: str
-    path: tuple[str | int, ...] = Field(repr=False)
-    kind: JsonNodeKind
-
-
-class JsonTextNode(StrictDomainModel):
-    """One selected JSON string with a stable document-local identity."""
-
-    id: str
-    path: tuple[str | int, ...] = Field(repr=False)
-    text: str = Field(repr=False)
 
 
 class JsonDocument:
@@ -202,23 +177,23 @@ class JsonDocument:
 
     def replace_text(
         self,
-        replacements: tuple[tuple[str, str], ...],
+        replacements: tuple[TextReplacement, ...],
         *,
         timeout: Timeout,
     ) -> bytes:
         """Replace complete JSON string tokens and preserve all unrelated source."""
         timeout.raise_if_expired()
-        replacement_ids = tuple(node_id for node_id, _ in replacements)
+        replacement_ids = tuple(item.target_id for item in replacements)
         if len(replacement_ids) != len(set(replacement_ids)):
             raise ValueError("JSON replacement node IDs must be unique")
         edits: list[tuple[int, int, bytes]] = []
-        for node_id, text in replacements:
+        for replacement in replacements:
             timeout.raise_if_expired()
-            node = self._nodes.get(node_id)
+            node = self._nodes.get(replacement.target_id)
             if node is None or node.kind is not JsonNodeKind.STRING:
                 raise ValueError("JSON replacement node ID is unknown")
             try:
-                rendered = json.dumps(text, ensure_ascii=False).encode(
+                rendered = json.dumps(replacement.text, ensure_ascii=False).encode(
                     "utf-8", errors="strict"
                 )
             except UnicodeEncodeError:
@@ -321,213 +296,6 @@ class JsonDocument:
         return public
 
 
-@dataclass(frozen=True)
-class _JsonNode:
-    id: str
-    path: tuple[str | int, ...]
-    kind: JsonNodeKind
-    token_start: int
-    token_end: int
-    text: str | None = None
-    object_members: tuple[tuple[str, _JsonNode], ...] = ()
-    array_items: tuple[_JsonNode, ...] = ()
-
-
-class _JsonParser:
-    def __init__(self, source: str, timeout: Timeout) -> None:
-        self.source = source
-        self.timeout = timeout
-        self.index = 0
-        self.node_count = 0
-        self.nodes: dict[str, _JsonNode] = {}
-
-    def parse(self) -> _JsonNode:
-        self._skip_whitespace()
-        root = self._parse_value((), 0)
-        self._skip_whitespace()
-        if self.index != len(self.source):
-            raise ValueError("trailing JSON data")
-        return root
-
-    def _parse_value(self, path: tuple[str | int, ...], depth: int) -> _JsonNode:
-        self.timeout.raise_if_expired()
-        if self.index >= len(self.source):
-            raise ValueError("JSON value is missing")
-        character = self.source[self.index]
-        if character == "{":
-            if depth >= MAX_JSON_DEPTH:
-                raise GateLimitExceededError("JSON nesting depth exceeds the limit")
-            return self._parse_object(path, depth)
-        if character == "[":
-            if depth >= MAX_JSON_DEPTH:
-                raise GateLimitExceededError("JSON nesting depth exceeds the limit")
-            return self._parse_array(path, depth)
-        if character == '"':
-            start = self.index
-            text = self._parse_string()
-            return self._new_node(
-                path,
-                JsonNodeKind.STRING,
-                start,
-                self.index,
-                text=text,
-            )
-        if character in "-0123456789":
-            start = self.index
-            match = _NUMBER_PATTERN.match(self.source, self.index)
-            if match is None:
-                raise ValueError("JSON number is invalid")
-            self.index = match.end()
-            return self._new_node(path, JsonNodeKind.NUMBER, start, self.index)
-        for literal, kind in (
-            ("true", JsonNodeKind.BOOLEAN),
-            ("false", JsonNodeKind.BOOLEAN),
-            ("null", JsonNodeKind.NULL),
-        ):
-            if self.source.startswith(literal, self.index):
-                start = self.index
-                self.index += len(literal)
-                return self._new_node(path, kind, start, self.index)
-        raise ValueError("JSON value is invalid")
-
-    def _parse_object(self, path: tuple[str | int, ...], depth: int) -> _JsonNode:
-        start = self.index
-        self.index += 1
-        self._skip_whitespace()
-        members: list[tuple[str, _JsonNode]] = []
-        keys: set[str] = set()
-        if self._consume("}"):
-            return self._new_node(
-                path, JsonNodeKind.OBJECT, start, self.index, object_members=()
-            )
-        while True:
-            if self.index >= len(self.source) or self.source[self.index] != '"':
-                raise ValueError("JSON object key is invalid")
-            key = self._parse_string()
-            if key in keys:
-                raise ValueError("JSON object keys must be unique")
-            keys.add(key)
-            self._skip_whitespace()
-            if not self._consume(":"):
-                raise ValueError("JSON object separator is missing")
-            self._skip_whitespace()
-            value = self._parse_value((*path, key), depth + 1)
-            members.append((key, value))
-            self._skip_whitespace()
-            if self._consume("}"):
-                break
-            if not self._consume(","):
-                raise ValueError("JSON object delimiter is missing")
-            self._skip_whitespace()
-        return self._new_node(
-            path,
-            JsonNodeKind.OBJECT,
-            start,
-            self.index,
-            object_members=tuple(members),
-        )
-
-    def _parse_array(self, path: tuple[str | int, ...], depth: int) -> _JsonNode:
-        start = self.index
-        self.index += 1
-        self._skip_whitespace()
-        items: list[_JsonNode] = []
-        if self._consume("]"):
-            return self._new_node(
-                path, JsonNodeKind.ARRAY, start, self.index, array_items=()
-            )
-        while True:
-            item = self._parse_value((*path, len(items)), depth + 1)
-            items.append(item)
-            self._skip_whitespace()
-            if self._consume("]"):
-                break
-            if not self._consume(","):
-                raise ValueError("JSON array delimiter is missing")
-            self._skip_whitespace()
-        return self._new_node(
-            path,
-            JsonNodeKind.ARRAY,
-            start,
-            self.index,
-            array_items=tuple(items),
-        )
-
-    def _parse_string(self) -> str:
-        start = self.index
-        self.index += 1
-        next_timeout_check = self.index + _TIMEOUT_CHECK_INTERVAL
-        while self.index < len(self.source):
-            if self.index >= next_timeout_check:
-                self.timeout.raise_if_expired()
-                next_timeout_check = self.index + _TIMEOUT_CHECK_INTERVAL
-            character = self.source[self.index]
-            if character == '"':
-                self.index += 1
-                token = self.source[start : self.index]
-                value = json.loads(token)
-                value.encode("utf-8", errors="strict")
-                return value
-            if character == "\\":
-                self.index += 1
-                if self.index >= len(self.source):
-                    raise ValueError("JSON string escape is incomplete")
-                escape = self.source[self.index]
-                if escape == "u":
-                    digits = self.source[self.index + 1 : self.index + 5]
-                    if len(digits) != 4 or _HEX_PATTERN.fullmatch(digits) is None:
-                        raise ValueError("JSON Unicode escape is invalid")
-                    self.index += 5
-                    continue
-                if escape not in '"\\/bfnrt':
-                    raise ValueError("JSON string escape is invalid")
-            elif ord(character) < 0x20:
-                raise ValueError("JSON string contains a control character")
-            self.index += 1
-        raise ValueError("JSON string is unterminated")
-
-    def _new_node(
-        self,
-        path: tuple[str | int, ...],
-        kind: JsonNodeKind,
-        token_start: int,
-        token_end: int,
-        *,
-        text: str | None = None,
-        object_members: tuple[tuple[str, _JsonNode], ...] = (),
-        array_items: tuple[_JsonNode, ...] = (),
-    ) -> _JsonNode:
-        self.node_count += 1
-        if self.node_count > MAX_JSON_NODES:
-            raise GateLimitExceededError("JSON node count exceeds the limit")
-        node = _JsonNode(
-            id=f"json-node-{self.node_count}",
-            path=path,
-            kind=kind,
-            token_start=token_start,
-            token_end=token_end,
-            text=text,
-            object_members=object_members,
-            array_items=array_items,
-        )
-        self.nodes[node.id] = node
-        return node
-
-    def _skip_whitespace(self) -> None:
-        next_timeout_check = self.index + _TIMEOUT_CHECK_INTERVAL
-        while self.index < len(self.source) and self.source[self.index] in " \t\r\n":
-            self.index += 1
-            if self.index >= next_timeout_check:
-                self.timeout.raise_if_expired()
-                next_timeout_check = self.index + _TIMEOUT_CHECK_INTERVAL
-
-    def _consume(self, token: str) -> bool:
-        if self.source.startswith(token, self.index):
-            self.index += len(token)
-            return True
-        return False
-
-
 def _select_segment(
     node: _JsonNode,
     segment: JsonPathSegment,
@@ -551,8 +319,6 @@ def _select_segment(
     return ()
 
 
-_NUMBER_PATTERN = re.compile(r"-?(?:0|[1-9][0-9]*)(?:\.[0-9]+)?(?:[eE][+-]?[0-9]+)?")
-_HEX_PATTERN = re.compile(r"[0-9a-fA-F]{4}")
 _TIMEOUT_CHECK_INTERVAL = 256
