+++ projects/egress-gate/src/egress_gate/request_content/text.py
@@ -1,0 +1,22 @@
+"""Shared immutable values for parsed request text and its replacements."""
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True)
+class TextTarget:
+    """One independently inspected text value and its content-local identity."""
+
+    id: str
+    text: str
+
+
+@dataclass(frozen=True, slots=True)
+class TextReplacement:
+    """The complete replacement text for one parsed target."""
+
+    target_id: str
+    text: str
+
+
+__all__ = ["TextReplacement", "TextTarget"]
